Add CRC32 with SSE support (#30)
This adds a CRC32 impl based on the SSE4.2 intrinsics. The hash_util::hash
function now decide at build time which hash to use. To enable the CRC32 hash,
one can pass the flags through RUSTC flag:

  RUSTFLAGS="-C target-feature=+sse4.2" cargo bench

Benchmarking shows that it can give some improvements.

Before:
test dict_i32_1k_10             ... bench:      13,915 ns/iter (+/- 313)
test dict_i32_1k_100            ... bench:      14,208 ns/iter (+/- 385)
test dict_i32_1k_1000           ... bench:      18,119 ns/iter (+/- 9,656)
test dict_i32_1m_10             ... bench:  14,113,301 ns/iter (+/- 272,263)
test dict_i32_1m_100            ... bench:  14,506,331 ns/iter (+/- 315,964)
test dict_i32_1m_1000           ... bench:  19,420,768 ns/iter (+/- 501,641)
test dict_str_1m                ... bench:  12,776,750 ns/iter (+/- 7,132,524)

After:
test dict_i32_1k_10             ... bench:      10,635 ns/iter (+/- 417)
test dict_i32_1k_100            ... bench:      10,778 ns/iter (+/- 133)
test dict_i32_1k_1000           ... bench:      16,894 ns/iter (+/- 223)
test dict_i32_1m_10             ... bench:  11,069,272 ns/iter (+/- 251,322)
test dict_i32_1m_100            ... bench:  11,080,385 ns/iter (+/- 215,262)
test dict_i32_1m_1000           ... bench:  11,445,189 ns/iter (+/- 163,009)
test dict_str_1m                ... bench:  12,963,771 ns/iter (+/- 2,493,590)

+++ src/encodings/encoding.rs
@@ -179,7 +179,7 @@ pub struct DictEncoder<T: DataType> {
 
   // Store `hash_table_size` - 1, so that `j & mod_bitmask` is equivalent to
   // `j % hash_table_size`, but uses far fewer CPU cycles.
-  mod_bitmask: u64,
+  mod_bitmask: u32,
 
   // Stores indices which map (many-to-one) to the values in the `uniques` array.
   // Here we are using fix-sized array with linear probing.
@@ -206,7 +206,7 @@ default impl<T: DataType> DictEncoder<T> {
     Self {
       desc: desc,
       hash_table_size: INITIAL_HASH_TABLE_SIZE,
-      mod_bitmask: (INITIAL_HASH_TABLE_SIZE - 1) as u64,
+      mod_bitmask: (INITIAL_HASH_TABLE_SIZE - 1) as u32,
       hash_slots: slots,
       buffered_indices: Buffer::new().with_mem_tracker(mem_tracker.clone()),
       uniques: Buffer::new().with_mem_tracker(mem_tracker.clone()),
@@ -306,7 +306,7 @@ default impl<T: DataType> DictEncoder<T> {
         continue;
       }
       let value = &self.uniques[index as usize];
-      let mut j = (hash_util::hash(value, 0) & ((new_size - 1) as u64)) as usize;
+      let mut j = (hash_util::hash(value, 0) & ((new_size - 1) as u32)) as usize;
       let mut slot = new_hash_slots[j];
       while slot != HASH_SLOT_EMPTY && self.uniques[slot as usize] != *value {
         j += 1;
@@ -320,7 +320,7 @@ default impl<T: DataType> DictEncoder<T> {
     }
 
     self.hash_table_size = new_size;
-    self.mod_bitmask = (new_size - 1) as u64;
+    self.mod_bitmask = (new_size - 1) as u32;
     mem::replace(&mut self.hash_slots, new_hash_slots);
   }
 }
@@ -511,7 +511,6 @@ default impl<T: DataType> Encoder<T> for DeltaBitPackEncoder<T> {
         self.flush_block_values()?;
       }
     }
-
     Ok(())
   }
